Good work!
The human can now interact with a tree and cut it down! When it has been cut down, it will disappear and the human will return to the house. Have some UI now which will let you know when you have correctly selected a human, need to add trees and other objects. There is HitPoints UI but it is very basic.

Very very spaghetti code right now, which is not great. But it currently works.

Things to do:
Improve UI
Look into refactoring early
Animations
Improve current assets