THE SPAGHETTI
So, I am not sure if things are getting better or worse lol.
Changed the Resource class to an `Abstract` class which I have never used before. I believe the benefit is that I can have empty functions?

I will continue to work on making things more modular which would be nice, but making good progress.

Implemented some nicer UI for when resources have been retrieved. Next it to use the retrieved resources to improve the Town Centre/Humans.

Looking good so far.