Code clean up and Uniqueness
Lots of clean up, using the private/public names of vars to reduce the number of functions and make things a bit easier to read.

Added a bit more uniqueness to the humans created. Their materials and icons will not have a different color assigned at creation to it is a bit easier to see which `Human` you have selected at any time.

I want to now work on the first improvement: movement speed! You will need 50 wood to increase the human move speed.